+++ src/core/extensions-api/has-default-export.function.ts
