+++ src/core/extensions-api/is-extension.function.ts
