+++ src/core/extensions-api/load-extension.function.ts
