+++ src/core/extensions-api/registry/extension.registry.test.ts
