+++ src/core/extensions-api/registry/extension.registry.ts
