+++ src/core/services/modal/layouts/property-editor-ui-picker/modal-layout-property-editor-ui-picker.element.ts
@@ -5,11 +5,11 @@ import { repeat } from 'lit/directives/repeat.js';
 import { Subscription } from 'rxjs';
 import { groupBy } from 'lodash';
 import type { UUIInputEvent } from '@umbraco-ui/uui';
-import { UmbContextConsumerMixin } from '../../../../context';
+import { UmbContextConsumerMixin } from '@umbraco-cms/context-api';
 
 import type { UmbModalHandler } from '../../modal-handler';
-import type { UmbExtensionRegistry } from '../../../../extension';
-import type { ManifestPropertyEditorUI } from '../../../../models';
+import type { UmbExtensionRegistry } from '@umbraco-cms/extensions-api';
+import type { ManifestPropertyEditorUI } from '@umbraco-cms/models';
 
 export interface UmbModalPropertyEditorUIPickerData {
 	selection?: Array<string>;
